*(ticdc): remove DEPRECATED sorters
