+++ cdc/capture/capture.go
@@ -28,7 +28,6 @@ import (
 	"github.com/pingcap/tiflow/cdc/owner"
 	"github.com/pingcap/tiflow/cdc/processor"
 	"github.com/pingcap/tiflow/cdc/processor/sourcemanager/engine/factory"
-	ssystem "github.com/pingcap/tiflow/cdc/sorter/db/system"
 	"github.com/pingcap/tiflow/pkg/config"
 	cdcContext "github.com/pingcap/tiflow/pkg/context"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
@@ -95,11 +94,6 @@ type captureImpl struct {
 	createEtcdClient createEtcdClientFunc
 	EtcdClient       etcd.CDCEtcdClient
 
-	// useSortEngine indicates whether to use the new pull based sort engine or
-	// the old push based sorter system. the latter will be removed after all sorter
-	// have been transformed into pull based sort engine.
-	useSortEngine     bool
-	sorterSystem      *ssystem.System
 	sortEngineFactory *factory.SortEngineFactory
 
 	// MessageServer is the receiver of the messages from the other nodes.
@@ -134,7 +128,6 @@ func NewCapture(pdEndpoints []string,
 	createEtcdClient createEtcdClientFunc,
 	grpcService *p2p.ServerWrapper,
 	sortEngineMangerFactory *factory.SortEngineFactory,
-	sorterSystem *ssystem.System,
 ) Capture {
 	return &captureImpl{
 		config:              config.GetGlobalServerConfig(),
@@ -147,10 +140,7 @@ func NewCapture(pdEndpoints []string,
 		newOwner:            owner.NewOwner,
 		info:                &model.CaptureInfo{},
 		createEtcdClient:    createEtcdClient,
-
-		useSortEngine:     sortEngineMangerFactory != nil,
-		sortEngineFactory: sortEngineMangerFactory,
-		sorterSystem:      sorterSystem,
+		sortEngineFactory:   sortEngineMangerFactory,
 	}
 }
 
@@ -328,7 +318,6 @@ func (c *captureImpl) run(stdCtx context.Context) error {
 		EtcdClient:        c.EtcdClient,
 		MessageServer:     c.MessageServer,
 		MessageRouter:     c.MessageRouter,
-		SorterSystem:      c.sorterSystem,
 		SortEngineFactory: c.sortEngineFactory,
 	})
