+++ cdc/processor/sourcemanager/engine/factory/pebble.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/cockroachdb/pebble"
 	"github.com/pingcap/log"
+	"github.com/pingcap/tiflow/cdc/processor/sourcemanager/engine"
 	epebble "github.com/pingcap/tiflow/cdc/processor/sourcemanager/engine/pebble"
-	metrics "github.com/pingcap/tiflow/cdc/sorter/db"
 	"github.com/pingcap/tiflow/pkg/config"
 	"go.uber.org/zap"
 )
@@ -52,7 +52,7 @@ func createPebbleDBs(
 			}
 			opts.EventListener.CompactionEnd = func(job pebble.CompactionInfo) {
 				idstr := strconv.Itoa(id + 1)
-				x := metrics.SorterCompactionDuration().WithLabelValues(idstr)
+				x := engine.SorterCompactionDuration().WithLabelValues(idstr)
 				x.Observe(job.TotalDuration.Seconds())
 			}
 		}
