+++ pkg/context/context.go
@@ -20,7 +20,6 @@ import (
 	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/cdc/processor/sourcemanager/engine/factory"
-	ssystem "github.com/pingcap/tiflow/cdc/sorter/db/system"
 	"github.com/pingcap/tiflow/pkg/config"
 	"github.com/pingcap/tiflow/pkg/etcd"
 	"github.com/pingcap/tiflow/pkg/p2p"
@@ -36,10 +35,6 @@ type GlobalVars struct {
 	EtcdClient  etcd.CDCEtcdClient
 
 	// SortEngineManager is introduced for pull-based sinks.
-	//
-	// TODO(qupeng): remove SorterSystem after all sorters are transformed
-	// to adapt pull-based sinks.
-	SorterSystem      *ssystem.System
 	SortEngineFactory *factory.SortEngineFactory
 
 	// OwnerRevision is the Etcd revision when the owner got elected.
