+++ pkg/db/db.go
@@ -27,7 +27,6 @@ type DB interface {
 	DeleteRange(start, end []byte) error
 	Compact(start, end []byte) error
 	Close() error
-	CollectMetrics(id int)
 }
 
 // A Batch is a sequence of Puts and Deletes that Commit to DB.
