+++ pkg/db/encoding/key_test.go
